Make website more responsibe and mobile-ready (#26)
Improve the look and feel of the website on mobile devices: wrap objects
in home screen, remove external links in the navbar, create a
collapsible menu for the navbar when the screen are too small and add a
burger menu to show/hide it. Disable pointer events on map, and add map
overlay that activates them after clicking on it.